Updated the methods for the ICO process.
Now, every function of both the bid and the crowdfunding processes gets the referendum's ID as byte[]. Moreover, the map of a member bid on a PP crowdfunding is made by a unique 'bidID' as a 256 hash value. However, I'm not sure if a function to get the value from a bid is created.